Rename channel states
