+++ packages/wallet/src/redux/states/channels/closingState.ts
@@ -4,7 +4,7 @@ import {
   TransactionExists,
   UserAddressExists,
   userAddressExists,
-} from './shared';
+} from './sharedState';
 
 // stage
 export const CLOSING = 'CLOSING';
