+++ packages/wallet/src/redux/states/channels/sharedState.ts
